unify async helpers as latest/coalesced/queued in src/lib/concurrency
One file, one naming family, same wrapper shape: take an async fn,
return a same-signature fn. Electron imports it like desktopApi/types.
