+++ apps/desktop/tsconfig.node.json
@@ -10,6 +10,7 @@
 		"vite.config.ts",
 		"electron.vite.config.ts",
 		"electron",
-		"src/desktopApi/types.ts"
+		"src/desktopApi/types.ts",
+		"src/lib/concurrency.ts"
 	]
 }
